Key location moved up level.
