+++ src/app.py
@@ -2,12 +2,11 @@ import json
 
 from flask import Flask, request, make_response, redirect, url_for
 
+from keys import VERIFICATION_TOKEN
 from src.route_decorators import validate_response, url_verification
 from src.routing_interface import combined_route_director
-from utils.config import get_instance_folder_path
 from utils import setup_logging
-from utils.keys import VERIFICATION_TOKEN
-
+from utils.file_config import get_instance_folder_path
 
 app = Flask(__name__,
             instance_path=get_instance_folder_path(),
